requirements.txt
django>=1.9.1
redis==2.10.5
requests>=2.9.1
django-debug-toolbar>=1.5
